Split out an effecthash crate
This needs to be in an independent crate at the bottom of the dep tree, so that
every component can depend on it and implement `EffectingData` to determine how
that componen's data should be effect-hashed.
