+++ crates/core/app/src/tests/spend.rs
@@ -5,8 +5,9 @@ use cnidarium::{ArcStateDeltaExt, StateDelta, TempStorage};
 use cnidarium_component::{ActionHandler as _, Component};
 use decaf377_rdsa::SigningKey;
 use penumbra_asset::Value;
-use penumbra_chain::{component::StateWriteExt, EffectHash, TransactionContext};
+use penumbra_chain::{component::StateWriteExt, TransactionContext};
 use penumbra_compact_block::component::CompactBlockManager;
+use penumbra_effecthash::EffectHash;
 use penumbra_keys::{test_keys, PayloadKey};
 use penumbra_num::Amount;
 use penumbra_sct::component::SourceContext;
