+++ crates/core/component/chain/Cargo.toml
@@ -11,6 +11,7 @@ penumbra-asset = { path = "../../../core/asset", default-features = false }
 penumbra-proto = { path = "../../../proto", default-features = false }
 penumbra-tct = { path = "../../../crypto/tct" }
 penumbra-keys = { path = "../../../core/keys", default-features = false }
+penumbra-effecthash = { path = "../../../core/effecthash", default-features = false }
 penumbra-num = { path = "../../../core/num", default-features = false }
 
 # Penumbra dependencies
