+++ crates/core/effecthash/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "penumbra-effecthash"
+version = "0.64.1"
+edition = "2021"
+
+[dependencies]
+# Workspace dependencies
+penumbra-proto = { path = "../../proto", default-features = false }
+anyhow = "1"
+hex = "0.4"
