+++ crates/core/effecthash/src/effecting_data.rs
@@ -1,0 +1,6 @@
+use crate::EffectHash;
+
+/// Something that can be hashed to produce an [`EffectHash`].
+pub trait EffectingData {
+    fn effect_hash(&self) -> EffectHash;
+}
