+++ crates/core/transaction/Cargo.toml
@@ -22,6 +22,7 @@ penumbra-fee = { path = "../component/fee/", default-features = false }
 penumbra-num = { path = "../num", default-features = false }
 penumbra-asset = { path = "../asset", default-features = false }
 penumbra-keys = { path = "../keys", default-features = false }
+penumbra-effecthash = { path = "../effecthash", default-features = false }
 
 # Git deps
 ibc-types = { version = "0.10.0", default-features = false }
