+++ crates/core/transaction/src/auth_data.rs
@@ -1,5 +1,5 @@
 use decaf377_rdsa::{Signature, SpendAuth};
-use penumbra_chain::EffectHash;
+use penumbra_effecthash::EffectHash;
 use penumbra_proto::{core::transaction::v1alpha1 as pb, DomainType};
 
 /// Authorization data returned in response to a
