+++ crates/custody/Cargo.toml
@@ -17,10 +17,11 @@ decaf377-frost = { path = "../crypto/decaf377-frost" }
 decaf377-ka = { path = "../crypto/decaf377-ka" }
 penumbra-chain = { path = "../core/component/chain" }
 penumbra-keys = { path = "../core/keys" }
+penumbra-effecthash = { path = "../core/effecthash" }
 penumbra-proto = { path = "../proto", features = ["rpc"] }
 penumbra-transaction = { path = "../core/transaction" }
 
-tokio = { version = "1.21.1", features = ["full"]}
+tokio = { version = "1.21.1", features = ["full"] }
 anyhow = "1"
 serde_json = "1"
 serde = { version = "1", features = ["derive"] }
