+++ crates/proto/src/lib.rs
@@ -61,6 +61,13 @@ pub mod penumbra {
             }
         }
 
+        pub mod effecthash {
+            pub mod v1alpha1 {
+                include!("gen/penumbra.core.effecthash.v1alpha1.rs");
+                include!("gen/penumbra.core.effecthash.v1alpha1.serde.rs");
+            }
+        }
+
         /// Components of the Penumbra application.
         pub mod component {
